# api/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AssumptionDistributionInput": {
        "properties": {
          "mean": {
            "description": "Fraction, e.g. 0.10 for 10%, not 10",
            "title": "Mean",
            "type": "number"
          },
          "std": {
            "description": "Standard deviation, same unit as mean",
            "minimum": 0.0,
            "title": "Std",
            "type": "number"
          }
        },
        "required": [
          "mean",
          "std"
        ],
        "title": "AssumptionDistributionInput",
        "type": "object"
      },
      "DataClass": {
        "enum": [
          "observed",
          "estimated",
          "simulated"
        ],
        "title": "DataClass",
        "type": "string"
      },
      "DebtLabScenarioBase": {
        "properties": {
          "divida_pib_pct": {
            "description": "Real observed value at reference_date",
            "title": "Divida Pib Pct",
            "type": "number"
          },
          "reference_date": {
            "description": "Most recent real divida_bruta_pib period used",
            "title": "Reference Date",
            "type": "string"
          },
          "source": {
            "description": "URL of the source series (BCB SGS 13762)",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "reference_date",
          "divida_pib_pct",
          "source"
        ],
        "title": "DebtLabScenarioBase",
        "type": "object"
      },
      "DebtLabScenarioRequest": {
        "properties": {
          "crescimento_nominal_pib": {
            "$ref": "#/components/schemas/AssumptionDistributionInput"
          },
          "horizon_years": {
            "default": 10,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Horizon Years",
            "type": "integer"
          },
          "juros_nominal": {
            "$ref": "#/components/schemas/AssumptionDistributionInput"
          },
          "n_iterations": {
            "default": 5000,
            "maximum": 20000.0,
            "minimum": 100.0,
            "title": "N Iterations",
            "type": "integer"
          },
          "primario_pct_pib": {
            "$ref": "#/components/schemas/AssumptionDistributionInput"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "juros_nominal",
          "crescimento_nominal_pib",
          "primario_pct_pib"
        ],
        "title": "DebtLabScenarioRequest",
        "type": "object"
      },
      "DebtLabScenarioResponse": {
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/DebtLabScenarioRequest"
          },
          "base": {
            "$ref": "#/components/schemas/DebtLabScenarioBase"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "data_class": {
            "$ref": "#/components/schemas/DataClass",
            "default": "simulated"
          },
          "deterministic_trajectory": {
            "items": {
              "$ref": "#/components/schemas/YearlyDeterministic"
            },
            "title": "Deterministic Trajectory",
            "type": "array"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "horizon_years": {
            "title": "Horizon Years",
            "type": "integer"
          },
          "n_iterations": {
            "title": "N Iterations",
            "type": "integer"
          },
          "percentiles": {
            "items": {
              "$ref": "#/components/schemas/YearlyPercentiles"
            },
            "title": "Percentiles",
            "type": "array"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "scenario_id",
          "created_at",
          "horizon_years",
          "base",
          "assumptions",
          "engine_version",
          "seed",
          "n_iterations",
          "deterministic_trajectory",
          "percentiles"
        ],
        "title": "DebtLabScenarioResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MetricResponse": {
        "properties": {
          "data_class": {
            "$ref": "#/components/schemas/DataClass"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "reference_date": {
            "title": "Reference Date",
            "type": "string"
          },
          "reference_year": {
            "title": "Reference Year",
            "type": "integer"
          },
          "state_ibge_code": {
            "title": "State Ibge Code",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "metric_id",
          "state_ibge_code",
          "value",
          "unit",
          "reference_year",
          "reference_date",
          "data_class",
          "provenance"
        ],
        "title": "MetricResponse",
        "type": "object"
      },
      "NationalMetricResponse": {
        "properties": {
          "data_class": {
            "$ref": "#/components/schemas/DataClass"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "reference_date": {
            "title": "Reference Date",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "metric_id",
          "value",
          "unit",
          "reference_date",
          "data_class",
          "provenance"
        ],
        "title": "NationalMetricResponse",
        "type": "object"
      },
      "ProvenanceResponse": {
        "properties": {
          "bronze_object": {
            "title": "Bronze Object",
            "type": "string"
          },
          "catalog_dataset_id": {
            "title": "Catalog Dataset Id",
            "type": "string"
          },
          "gold_object": {
            "title": "Gold Object",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "producing_organization": {
            "title": "Producing Organization",
            "type": "string"
          },
          "reference_date": {
            "title": "Reference Date",
            "type": "string"
          },
          "reference_year": {
            "title": "Reference Year",
            "type": "integer"
          },
          "silver_transform": {
            "title": "Silver Transform",
            "type": "string"
          },
          "silver_transform_version": {
            "title": "Silver Transform Version",
            "type": "string"
          },
          "source_resource_url": {
            "title": "Source Resource Url",
            "type": "string"
          },
          "state_ibge_code": {
            "title": "State Ibge Code",
            "type": "string"
          },
          "trust_status": {
            "title": "Trust Status",
            "type": "string"
          }
        },
        "required": [
          "metric_id",
          "state_ibge_code",
          "reference_year",
          "reference_date",
          "gold_object",
          "silver_transform",
          "silver_transform_version",
          "bronze_object",
          "source_resource_url",
          "catalog_dataset_id",
          "producing_organization",
          "trust_status"
        ],
        "title": "ProvenanceResponse",
        "type": "object"
      },
      "ProvenanceSummary": {
        "properties": {
          "reference_date": {
            "title": "Reference Date",
            "type": "string"
          },
          "source": {
            "description": "URL of the source resource",
            "title": "Source",
            "type": "string"
          },
          "trust_status": {
            "description": "Provenance trust status; 'source_only' until a Data Trust Score exists (SPEC-006)",
            "title": "Trust Status",
            "type": "string"
          }
        },
        "required": [
          "source",
          "reference_date",
          "trust_status"
        ],
        "title": "ProvenanceSummary",
        "type": "object"
      },
      "SuggestedAssumption": {
        "properties": {
          "mean": {
            "description": "Fraction, e.g. 0.10 for 10%, not 10",
            "title": "Mean",
            "type": "number"
          },
          "methodology": {
            "description": "Plain-text note on how mean/std were derived from the real series",
            "title": "Methodology",
            "type": "string"
          },
          "period_end": {
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "title": "Period Start",
            "type": "string"
          },
          "source_metric_id": {
            "description": "Gold metric_id the suggestion was derived from",
            "title": "Source Metric Id",
            "type": "string"
          },
          "std": {
            "description": "Standard deviation, same unit as mean",
            "minimum": 0.0,
            "title": "Std",
            "type": "number"
          },
          "window_months": {
            "description": "Number of real trailing months used",
            "title": "Window Months",
            "type": "integer"
          }
        },
        "required": [
          "mean",
          "std",
          "window_months",
          "source_metric_id",
          "period_start",
          "period_end",
          "methodology"
        ],
        "title": "SuggestedAssumption",
        "type": "object"
      },
      "SuggestedAssumptionsResponse": {
        "properties": {
          "crescimento_nominal_pib": {
            "$ref": "#/components/schemas/SuggestedAssumption"
          },
          "data_class": {
            "$ref": "#/components/schemas/DataClass",
            "default": "estimated"
          },
          "juros_nominal": {
            "$ref": "#/components/schemas/SuggestedAssumption"
          }
        },
        "required": [
          "juros_nominal",
          "crescimento_nominal_pib"
        ],
        "title": "SuggestedAssumptionsResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "YearlyDeterministic": {
        "properties": {
          "divida_pib_pct": {
            "title": "Divida Pib Pct",
            "type": "number"
          },
          "year_offset": {
            "title": "Year Offset",
            "type": "integer"
          }
        },
        "required": [
          "year_offset",
          "divida_pib_pct"
        ],
        "title": "YearlyDeterministic",
        "type": "object"
      },
      "YearlyPercentiles": {
        "properties": {
          "p10": {
            "title": "P10",
            "type": "number"
          },
          "p25": {
            "title": "P25",
            "type": "number"
          },
          "p50": {
            "title": "P50",
            "type": "number"
          },
          "p75": {
            "title": "P75",
            "type": "number"
          },
          "p90": {
            "title": "P90",
            "type": "number"
          },
          "year_offset": {
            "title": "Year Offset",
            "type": "integer"
          }
        },
        "required": [
          "year_offset",
          "p10",
          "p25",
          "p50",
          "p75",
          "p90"
        ],
        "title": "YearlyPercentiles",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Walking skeleton: consolidated state debt with full provenance (SPEC-033).",
    "title": "BRASIL 2036 \u2014 Metrics API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    },
    "/v1/metrics/{metric_id}": {
      "get": {
        "operationId": "get_metric_v1_metrics__metric_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "metric_id",
            "required": true,
            "schema": {
              "title": "Metric Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state_ibge_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Ibge Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Metric"
      }
    },
    "/v1/metrics/{metric_id}/national": {
      "get": {
        "operationId": "get_national_metric_v1_metrics__metric_id__national_get",
        "parameters": [
          {
            "in": "path",
            "name": "metric_id",
            "required": true,
            "schema": {
              "title": "Metric Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NationalMetricResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get National Metric"
      }
    },
    "/v1/provenance/{metric_id}": {
      "get": {
        "operationId": "get_provenance_v1_provenance__metric_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "metric_id",
            "required": true,
            "schema": {
              "title": "Metric Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state_ibge_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Ibge Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provenance"
      }
    },
    "/v1/simulations/debtlab": {
      "post": {
        "operationId": "create_debtlab_scenario_v1_simulations_debtlab_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebtLabScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtLabScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Debtlab Scenario"
      }
    },
    "/v1/simulations/debtlab/suggested-assumptions": {
      "get": {
        "operationId": "get_suggested_assumptions_v1_simulations_debtlab_suggested_assumptions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestedAssumptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Suggested Assumptions"
      }
    },
    "/v1/simulations/debtlab/{scenario_id}": {
      "get": {
        "operationId": "get_debtlab_scenario_v1_simulations_debtlab__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtLabScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Debtlab Scenario"
      }
    }
  }
}